+++ code/cell.modules/pkg.sys/ui.dev/src/types/index.ts
@@ -1,4 +1,4 @@
 export * from './types';
-export * from './types.Event';
+export * from './types.event';
 export * from './types.Host';
 export * from './types.Actions';
