+++ code/cell.modules/pkg.sys/ui.dev/src/types/types.Actions/Actions.item.Button.ts
@@ -1,0 +1,18 @@
+import { t } from '../common';
+
+/**
+ * INPUT: Simple clickable action.
+ */
+export type DevActionButton = t.DevActionButtonProps & {
+  id: string;
+  kind: 'button';
+  handler?: t.DevActionButtonHandler<any>;
+};
+
+/**
+ * Editable properties of a [Button].
+ */
+export type DevActionButtonProps = {
+  label: string;
+  description?: string;
+};
